Cambios en producto y unidades

+++ t3-ej1/src/clases/Granel.java
@@ -2,13 +2,11 @@ package clases;
 
 public class Granel extends Producto {
 
-	public float cantidad;
+	public double cantidad;
 	
-	public Granel(int id, String nombre, String category, float precio, float cantidad) {
+	public Granel(int id, String nombre, String category, double precio, double cantidad) {
 		super(id, nombre, category, precio);
 		this.cantidad = cantidad;
 	}
-	
-	
-	
+		
 }
