+++ t3-ej1/src/clases/Producto.java
@@ -5,15 +5,13 @@ public class Producto {
 	public int id;
 	public String nombre;
 	public String category;
-	public float precio;
+	public double precio;
 	
-	public Producto(int id, String nombre, String category, float precio) {
+	public Producto(int id, String nombre, String category, double precio) {
 		this.id = id;
 		this.nombre = nombre;
 		this.category = category;
 		this.precio = precio;
 	}
 	
-	
-
 }
