+++ t3-ej1/src/clases/Unidades.java
@@ -4,9 +4,9 @@ public class Unidades extends Producto {
 	
 	public int unidades;
 
-	public Unidades(int id, String nombre, String category, float precio, int unidades) {
+	public Unidades(int id, String nombre, String category, double precio, int unidades) {
 		super(id, nombre, category, precio);
 		this.unidades = unidades;
 	}
-
+	
 }
